moved layout struct
